+++ NEWS.md
@@ -1,7 +1,30 @@
+# pctax v0.1.6 Notes
+
+## Fixed
+
+- fixed `taxonkit` function for windows <2025-06-12, Thu>
+
+# pctax v0.1.5 Notes
+
+## Added
+
+- added `load_mpa_df` function <2025-06-12, Thu>
+- added `pre_assembly_stats` function <2025-06-12, Thu>
+
+# pctax v0.1.4 Notes
+
+## Added
+
+- added `stamp_plot` function <2025-06-12, Thu>
+- added `multi_conf` function <2025-06-12, Thu>
+
 # pctax v0.1.3 Notes
 
 ## Added
 
+- added mode 4 for `plot.b_res` function <2025-04-27, Sun>
+- added `drop_tips_update()` function <2025-04-27, Sun>
+- added `stamp_plot` function <2025-01-14, Tue>
 - added `convert_taxon_name` function <2024-11-19, Tue>
 - added `plot_two_tree` function <2024-10-22, Tue>